+++ example/lib/main.dart
@@ -12,6 +12,7 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:pspdfkit_example/pspdfkit_save_as_example.dart';
 
 import 'package:pspdfkit_flutter/src/main.dart';
 import 'package:pspdfkit_flutter/src/widgets/pspdfkit_widget_controller.dart';
@@ -20,7 +21,10 @@ import 'package:pspdfkit_flutter/src/widgets/pspdfkit_widget.dart';
 import 'pspdfkit_form_example.dart';
 import 'pspdfkit_instantjson_example.dart';
 import 'pspdfkit_annotations_example.dart';
+import 'pspdfkit_manual_save_example.dart';
+import 'pspdfkit_save_as_example.dart';
 import 'pspdfkit_annotation_processing_example.dart';
+import 'platform_utils.dart';
 
 const String _documentPath = 'PDFs/PSPDFKit.pdf';
 const String _lockedDocumentPath = 'PDFs/protected.pdf';
@@ -57,6 +61,12 @@ const String _annotationsExample =
     'Programmatically Adds and Removes Annotations';
 const String _annotationsExampleSub =
     'Programmatically adds and removes annotations using a custom Widget.';
+const String _manualSaveExample = 'Manual Save';
+const String _saveAsExample = 'Save As';
+const String _manualSaveExampleSub =
+    'Add a save button at the bottom and disable automatic saving.';
+const String _saveAsExampleSub =
+    'Embed and save the changes made to a document into a new file';
 const String _annotationProcessingExample = 'Process Annotations';
 const String _annotationProcessingExampleSub =
     'Programmatically adds and removes annotations using a custom Widget.';
@@ -125,43 +135,34 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   String _frameworkVersion = '';
   ThemeData currentTheme = lightTheme;
 
-  Future<File> extractAsset(String assetPath) async {
+  Future<File> extractAsset(String assetPath,
+      {bool shouldOverwrite = true, String prefix = ''}) async {
     final bytes = await DefaultAssetBundle.of(context).load(assetPath);
     final list = bytes.buffer.asUint8List();
 
     final tempDir = await Pspdfkit.getTemporaryDirectory();
-    final tempDocumentPath = '${tempDir.path}/$assetPath';
+    final tempDocumentPath = '${tempDir.path}/$prefix$assetPath';
+    final file = File(tempDocumentPath);
 
-    final file = await File(tempDocumentPath).create(recursive: true);
-    file.writeAsBytesSync(list);
-    return file;
-  }
-
-  bool isCupertino(BuildContext context) {
-    final defaultTargetPlatform = Theme.of(context).platform;
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        return true;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        return false;
+    if (shouldOverwrite || !file.existsSync()) {
+      await file.create(recursive: true);
+      file.writeAsBytesSync(list);
     }
+    return file;
   }
 
   void showDocument() async {
     final extractedDocument = await extractAsset(_documentPath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
         builder: (_) => Scaffold(
-            extendBodyBehindAppBar: isCupertino(context) ? false : true,
+            extendBodyBehindAppBar:
+                PlatformUtils.isCupertino(context) ? false : true,
             appBar: AppBar(),
             body: SafeArea(
                 top: false,
                 bottom: false,
                 child: Container(
-                    padding: isCupertino(context)
+                    padding: PlatformUtils.isCupertino(context)
                         ? null
                         : const EdgeInsets.only(top: kToolbarHeight),
                     child: PspdfkitWidget(
@@ -171,14 +172,14 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   void showDocumentPlatformStyle() async {
     final extractedDocument = await extractAsset(_documentPath);
 
-    if (isCupertino(context)) {
+    if (PlatformUtils.isCupertino(context)) {
       await Navigator.of(context).push<dynamic>(CupertinoPageRoute<dynamic>(
           builder: (_) => CupertinoPageScaffold(
               navigationBar: const CupertinoNavigationBar(),
               child: SafeArea(
                   bottom: false,
                   child:
-                  PspdfkitWidget(documentPath: extractedDocument.path)))));
+                      PspdfkitWidget(documentPath: extractedDocument.path)))));
     } else {
       await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
           builder: (_) => Scaffold(
@@ -198,30 +199,32 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     final extractedImage = await extractAsset(_imagePath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
         builder: (_) => Scaffold(
-            extendBodyBehindAppBar: isCupertino(context) ? false : true,
+            extendBodyBehindAppBar:
+                PlatformUtils.isCupertino(context) ? false : true,
             appBar: AppBar(),
             body: SafeArea(
                 top: false,
                 bottom: false,
                 child: Container(
-                    padding: isCupertino(context)
+                    padding: PlatformUtils.isCupertino(context)
                         ? null
                         : const EdgeInsets.only(top: kToolbarHeight),
                     child:
-                    PspdfkitWidget(documentPath: extractedImage.path))))));
+                        PspdfkitWidget(documentPath: extractedImage.path))))));
   }
 
   void applyDarkTheme() async {
     final extractedDocument = await extractAsset(_documentPath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
         builder: (_) => Scaffold(
-            extendBodyBehindAppBar: isCupertino(context) ? false : true,
+            extendBodyBehindAppBar:
+                PlatformUtils.isCupertino(context) ? false : true,
             appBar: AppBar(),
             body: SafeArea(
                 top: false,
                 bottom: false,
                 child: Container(
-                    padding: isCupertino(context)
+                    padding: PlatformUtils.isCupertino(context)
                         ? null
                         : const EdgeInsets.only(top: kToolbarHeight),
                     child: PspdfkitWidget(
@@ -229,7 +232,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                         configuration: const {
                           appearanceMode: appearanceModeNight,
                           androidDarkThemeResource:
-                          'PSPDFKit.Theme.Example.Dark'
+                              'PSPDFKit.Theme.Example.Dark'
                         }))))));
   }
 
@@ -237,25 +240,19 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     final extractedDocument = await extractAsset(_documentPath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
         builder: (_) => Scaffold(
-            extendBodyBehindAppBar: isCupertino(context) ? false : true,
+            extendBodyBehindAppBar:
+                PlatformUtils.isCupertino(context) ? false : true,
             appBar: AppBar(),
             body: SafeArea(
                 top: false,
                 bottom: false,
                 child: Container(
-                    padding: isCupertino(context)
+                    padding: PlatformUtils.isCupertino(context)
                         ? null
                         : const EdgeInsets.only(top: kToolbarHeight),
                     child: PspdfkitWidget(
                         documentPath: extractedDocument.path,
                         configuration: const {
-
-                          creatorName : "Alberto Ciaone",
-                          fullname: "Alberto Marcone",
-                          watermarkEnabled: true,
-                          watermarkColor: "#ff00ff",
-                          watermarkOpacity: 0.6,
-                          watermarkFontSize: 30,
                           scrollDirection: 'vertical',
                           pageTransition: 'scrollContinuous',
                           spreadFitting: 'fit',
@@ -287,9 +284,6 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                             'searchButtonItem',
                             'annotationButtonItem'
                           ],
-                          editableAnnotationTypes : [
-                            'PSPDFAnnotationStringHighlight'
-                          ],
                           iOSLeftBarButtonItems: ['settingsButtonItem'],
                           iOSAllowToolbarTitleChange: false,
                           toolbarTitle: 'Custom Title',
@@ -314,13 +308,14 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
     final extractedLockedDocument = await extractAsset(_lockedDocumentPath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
         builder: (_) => Scaffold(
-            extendBodyBehindAppBar: isCupertino(context) ? false : true,
+            extendBodyBehindAppBar:
+                PlatformUtils.isCupertino(context) ? false : true,
             appBar: AppBar(),
             body: SafeArea(
                 top: false,
                 bottom: false,
                 child: Container(
-                    padding: isCupertino(context)
+                    padding: PlatformUtils.isCupertino(context)
                         ? null
                         : const EdgeInsets.only(top: kToolbarHeight),
                     child: PspdfkitWidget(
@@ -351,6 +346,30 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
             documentPath: extractedDocument.path)));
   }
 
+  void manualSaveExample() async {
+    final extractedWritableDocument = await extractAsset(_documentPath,
+        shouldOverwrite: false, prefix: 'persist');
+
+    // Automatic Saving of documents is enabled by default in certain scenarios [see for details: https://pspdfkit.com/guides/flutter/save-a-document/#auto-save]
+    // In order to manually save documents, you might consider disabling automatic saving with disableAutosave: true in the config
+    await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
+        builder: (_) => PspdfkitManualSaveExampleWidget(
+            documentPath: extractedWritableDocument.path,
+            configuration: const {disableAutosave: true})));
+  }
+
+  void saveAsExample() async {
+    final extractedWritableDocument = await extractAsset(_documentPath,
+        shouldOverwrite: false, prefix: 'persist');
+
+    // Automatic Saving of documents is enabled by default in certain scenarios [see for details: https://pspdfkit.com/guides/flutter/save-a-document/#auto-save]
+    // In order to manually save documents, you might consider disabling automatic saving with disableAutosave: true in the config
+    await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
+        builder: (_) => PspdfkitSaveAsExampleWidget(
+            documentPath: extractedWritableDocument.path,
+            configuration: const {disableAutosave: true})));
+  }
+
   void annotationProcessingExample() async {
     final extractedDocument = await extractAsset(_documentPath);
     await Navigator.of(context).push<dynamic>(MaterialPageRoute<dynamic>(
@@ -364,7 +383,7 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
       final extractedDocument = await extractAsset(_documentPath);
       final extractedFormDocument = await extractAsset(_formPath);
 
-      // if (isCupertino(context)) {
+      if (PlatformUtils.isCupertino(context)) {
         await Navigator.of(context).push<dynamic>(CupertinoPageRoute<dynamic>(
             builder: (_) => CupertinoPageScaffold(
                 navigationBar: CupertinoNavigationBar(),
@@ -379,10 +398,10 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
                               documentPath: extractedFormDocument.path,
                               onPspdfkitWidgetCreated: onWidgetCreated))
                     ])))));
-     /* } else {
+      } else {
         // This example is only supported in iOS at the moment.
         // Support for Android is coming soon.
-      }*/
+      }
     } on PlatformException catch (e) {
       print("Failed to present document: '${e.message}'.");
     }
@@ -549,9 +568,9 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
   @override
   void didChangePlatformBrightness() {
     currentTheme =
-    WidgetsBinding.instance?.window.platformBrightness == Brightness.light
-        ? lightTheme
-        : darkTheme;
+        WidgetsBinding.instance?.window.platformBrightness == Brightness.light
+            ? lightTheme
+            : darkTheme;
     setState(() {
       build(context);
     });
@@ -655,20 +674,29 @@ class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
           title: const Text(_annotationsExample),
           subtitle: const Text(_annotationsExampleSub),
           onTap: () => annotationsExample()),
+      ListTile(
+          title: const Text(_manualSaveExample),
+          subtitle: const Text(_manualSaveExampleSub),
+          onTap: () => manualSaveExample()),
+      if (PlatformUtils.isCupertino(context))
+        ListTile(
+            title: const Text(_saveAsExample),
+            subtitle: const Text(_saveAsExampleSub),
+            onTap: () => saveAsExample()),
       // The annotation processing example is supported by iOS only for now.
-      if (isCupertino(context))
+      if (PlatformUtils.isCupertino(context))
         ListTile(
             title: const Text(_annotationProcessingExample),
             subtitle: const Text(_annotationProcessingExampleSub),
             onTap: () => annotationProcessingExample()),
       // The import Instant JSON example is supported by iOS only for now.
-      if (isCupertino(context))
+      if (PlatformUtils.isCupertino(context))
         ListTile(
             title: const Text(_importInstantJsonExample),
             subtitle: const Text(_importInstantJsonExampleSub),
             onTap: () => importInstantJsonExample()),
       // The push two PspdfWidgets simultaneously example is supported by iOS only for now.
-      // if (isCupertino(context))
+      if (PlatformUtils.isCupertino(context))
         ListTile(
             title: const Text(_widgetExampleFullScreen),
             subtitle: const Text(_widgetExampleFullScreenSub),