+++ example/lib/pspdfkit_manual_save_example.dart
@@ -1,5 +1,5 @@
 ///
-///  Copyright © 2018-2022 PSPDFKit GmbH. All rights reserved.
+///  Copyright © 2022 PSPDFKit GmbH. All rights reserved.
 ///
 ///  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
 ///  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
@@ -15,38 +15,30 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:pspdfkit_example/platform_utils.dart';
 
 import 'package:pspdfkit_flutter/src/widgets/pspdfkit_widget_controller.dart';
 
-typedef PspdfkitFormExampleWidgetCreatedCallback = void Function(
-    PspdfkitWidgetController view);
+import 'platform_utils.dart';
 
-class PspdfkitFormExampleWidget extends StatefulWidget {
+class PspdfkitManualSaveExampleWidget extends StatefulWidget {
   final String documentPath;
   final dynamic configuration;
-  final PspdfkitFormExampleWidgetCreatedCallback?
-      onPspdfkitFormExampleWidgetCreated;
 
-  const PspdfkitFormExampleWidget(
-      {Key? key,
-      required this.documentPath,
-      this.configuration,
-      this.onPspdfkitFormExampleWidgetCreated})
+  const PspdfkitManualSaveExampleWidget(
+      {Key? key, required this.documentPath, this.configuration})
       : super(key: key);
 
   @override
-  _PspdfkitFormExampleWidgetState createState() =>
-      _PspdfkitFormExampleWidgetState();
+  _PspdfkitManualSaveExampleWidgetState createState() =>
+      _PspdfkitManualSaveExampleWidgetState();
 }
 
-class _PspdfkitFormExampleWidgetState extends State<PspdfkitFormExampleWidget> {
-  late PspdfkitWidgetController view;
-  bool _keyboardVisible = false;
+class _PspdfkitManualSaveExampleWidgetState
+    extends State<PspdfkitManualSaveExampleWidget> {
+  late PspdfkitWidgetController pspdfkitWidgetController;
 
   @override
   Widget build(BuildContext context) {
-    _keyboardVisible = MediaQuery.of(context).viewInsets.bottom != 0;
     // This is used in the platform side to register the view.
     const String viewType = 'com.pspdfkit.widget';
     // Pass parameters to the platform side.
@@ -57,9 +49,6 @@ class _PspdfkitFormExampleWidgetState extends State<PspdfkitFormExampleWidget> {
     if (PlatformUtils.isCurrentPlatformSupported()) {
       return Scaffold(
           extendBodyBehindAppBar: PlatformUtils.isAndroid(),
-          // Do not resize the the document view on Android or
-          // it won't be rendered correctly when filling forms.
-          resizeToAvoidBottomInset: PlatformUtils.isIOS(),
           appBar: AppBar(),
           body: SafeArea(
               top: false,
@@ -104,8 +93,7 @@ class _PspdfkitFormExampleWidgetState extends State<PspdfkitFormExampleWidget> {
                                     ..addOnPlatformViewCreatedListener(
                                         onPlatformViewCreated)
                                     ..create();
-                                },
-                              )
+                                })
                             : UiKitView(
                                 viewType: viewType,
                                 layoutDirection: TextDirection.ltr,
@@ -113,43 +101,14 @@ class _PspdfkitFormExampleWidgetState extends State<PspdfkitFormExampleWidget> {
                                 onPlatformViewCreated: onPlatformViewCreated,
                                 creationParamsCodec:
                                     const StandardMessageCodec())),
-                    // On Android do not show the buttons when the Keyboard
-                    // is visible. PSPDFKit for Android automatically
-                    // fills the space available and re-render the document view.
-                    if (!_keyboardVisible || PlatformUtils.isIOS())
-                      SizedBox(
-                          child: Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                              children: <Widget>[
-                            ElevatedButton(
-                                onPressed: () {
-                                  view.setFormFieldValue(
-                                      'Updated Form Field Value', 'Name_Last');
-                                },
-                                child: const Text('Set form field value')),
-                            ElevatedButton(
-                                onPressed: () async {
-                                  const title = 'Form Field Value';
-                                  final formFieldValue = await view
-                                          .getFormFieldValue('Name_Last') ??
-                                      '';
-                                  await showDialog<AlertDialog>(
-                                      context: context,
-                                      builder: (BuildContext context) =>
-                                          AlertDialog(
-                                            title: const Text(title),
-                                            content: Text(formFieldValue),
-                                            actions: [
-                                              TextButton(
-                                                  onPressed: () {
-                                                    Navigator.of(context).pop();
-                                                  },
-                                                  child: const Text('OK'))
-                                            ],
-                                          ));
-                                },
-                                child: const Text('Get form field value'))
-                          ]))
+                    SizedBox(
+                        child: Column(children: <Widget>[
+                      ElevatedButton(
+                          onPressed: () async {
+                            await pspdfkitWidgetController.save();
+                          },
+                          child: const Text('Save Document'))
+                    ]))
                   ]))));
     } else {
       return Text(
@@ -157,14 +116,7 @@ class _PspdfkitFormExampleWidgetState extends State<PspdfkitFormExampleWidget> {
     }
   }
 
-  bool isKeyboardVisible() {
-    return MediaQuery.of(context).viewInsets.bottom == 0;
-  }
-
   Future<void> onPlatformViewCreated(int id) async {
-    view = PspdfkitWidgetController(id);
-    if (widget.onPspdfkitFormExampleWidgetCreated != null) {
-      widget.onPspdfkitFormExampleWidgetCreated!(view);
-    }
+    pspdfkitWidgetController = PspdfkitWidgetController(id);
   }
 }